HOTT-2128 Namespace News Items
We'll be adding a News::Collections model so will allow grouping related
these models together
